+++ packages/primeng/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/ngx-form-engine-primeng.component';
 export * from './lib/ngx-json-form-stepper.component';
 export * from './lib/ngx-json-form-debug.component';
+export * from './lib/signature-pad.component';